Move Component to own file
